+++ autoarray/plot/__init__.py
@@ -1,40 +1,40 @@
-from autoarray.plot.mat_wrap.wrap.wrap_base import Units
-from autoarray.plot.mat_wrap.wrap.wrap_base import Figure
-from autoarray.plot.mat_wrap.wrap.wrap_base import Axis
-from autoarray.plot.mat_wrap.wrap.wrap_base import Cmap
-from autoarray.plot.mat_wrap.wrap.wrap_base import Colorbar
-from autoarray.plot.mat_wrap.wrap.wrap_base import ColorbarTickParams
-from autoarray.plot.mat_wrap.wrap.wrap_base import TickParams
-from autoarray.plot.mat_wrap.wrap.wrap_base import YTicks
-from autoarray.plot.mat_wrap.wrap.wrap_base import XTicks
-from autoarray.plot.mat_wrap.wrap.wrap_base import Title
-from autoarray.plot.mat_wrap.wrap.wrap_base import YLabel
-from autoarray.plot.mat_wrap.wrap.wrap_base import XLabel
-from autoarray.plot.mat_wrap.wrap.wrap_base import Text
-from autoarray.plot.mat_wrap.wrap.wrap_base import Legend
-from autoarray.plot.mat_wrap.wrap.wrap_base import Output
+from autoarray.plot.wrap.wrap_base import Units
+from autoarray.plot.wrap.wrap_base import Figure
+from autoarray.plot.wrap.wrap_base import Axis
+from autoarray.plot.wrap.wrap_base import Cmap
+from autoarray.plot.wrap.wrap_base import Colorbar
+from autoarray.plot.wrap.wrap_base import ColorbarTickParams
+from autoarray.plot.wrap.wrap_base import TickParams
+from autoarray.plot.wrap.wrap_base import YTicks
+from autoarray.plot.wrap.wrap_base import XTicks
+from autoarray.plot.wrap.wrap_base import Title
+from autoarray.plot.wrap.wrap_base import YLabel
+from autoarray.plot.wrap.wrap_base import XLabel
+from autoarray.plot.wrap.wrap_base import Text
+from autoarray.plot.wrap.wrap_base import Legend
+from autoarray.plot.wrap.wrap_base import Output
 
-from autoarray.plot.mat_wrap.wrap.wrap_1d import YXPlot
-from autoarray.plot.mat_wrap.wrap.wrap_1d import YXScatter
-from autoarray.plot.mat_wrap.wrap.wrap_1d import AXVLine
-from autoarray.plot.mat_wrap.wrap.wrap_1d import FillBetween
+from autoarray.plot.wrap.wrap_1d import YXPlot
+from autoarray.plot.wrap.wrap_1d import YXScatter
+from autoarray.plot.wrap.wrap_1d import AXVLine
+from autoarray.plot.wrap.wrap_1d import FillBetween
 
-from autoarray.plot.mat_wrap.wrap.wrap_2d import ArrayOverlay
-from autoarray.plot.mat_wrap.wrap.wrap_2d import GridScatter
-from autoarray.plot.mat_wrap.wrap.wrap_2d import GridPlot
-from autoarray.plot.mat_wrap.wrap.wrap_2d import GridErrorbar
-from autoarray.plot.mat_wrap.wrap.wrap_2d import VectorFieldQuiver
-from autoarray.plot.mat_wrap.wrap.wrap_2d import PatchOverlay
-from autoarray.plot.mat_wrap.wrap.wrap_2d import VoronoiDrawer
-from autoarray.plot.mat_wrap.wrap.wrap_2d import OriginScatter
-from autoarray.plot.mat_wrap.wrap.wrap_2d import MaskScatter
-from autoarray.plot.mat_wrap.wrap.wrap_2d import BorderScatter
-from autoarray.plot.mat_wrap.wrap.wrap_2d import PositionsScatter
-from autoarray.plot.mat_wrap.wrap.wrap_2d import IndexScatter
-from autoarray.plot.mat_wrap.wrap.wrap_2d import PixelizationGridScatter
-from autoarray.plot.mat_wrap.wrap.wrap_2d import ParallelOverscanPlot
-from autoarray.plot.mat_wrap.wrap.wrap_2d import SerialPrescanPlot
-from autoarray.plot.mat_wrap.wrap.wrap_2d import SerialOverscanPlot
+from autoarray.plot.wrap.wrap_2d import ArrayOverlay
+from autoarray.plot.wrap.wrap_2d import GridScatter
+from autoarray.plot.wrap.wrap_2d import GridPlot
+from autoarray.plot.wrap.wrap_2d import GridErrorbar
+from autoarray.plot.wrap.wrap_2d import VectorFieldQuiver
+from autoarray.plot.wrap.wrap_2d import PatchOverlay
+from autoarray.plot.wrap.wrap_2d import VoronoiDrawer
+from autoarray.plot.wrap.wrap_2d import OriginScatter
+from autoarray.plot.wrap.wrap_2d import MaskScatter
+from autoarray.plot.wrap.wrap_2d import BorderScatter
+from autoarray.plot.wrap.wrap_2d import PositionsScatter
+from autoarray.plot.wrap.wrap_2d import IndexScatter
+from autoarray.plot.wrap.wrap_2d import PixelizationGridScatter
+from autoarray.plot.wrap.wrap_2d import ParallelOverscanPlot
+from autoarray.plot.wrap.wrap_2d import SerialPrescanPlot
+from autoarray.plot.wrap.wrap_2d import SerialOverscanPlot
 
 from autoarray.plot.mat_wrap.mat_plot import MatPlot1D
 from autoarray.plot.mat_wrap.include import Include1D
@@ -44,15 +44,15 @@ from autoarray.plot.mat_wrap.include import Include2D
 from autoarray.plot.mat_wrap.visuals import Visuals2D
 from autoarray.plot.mat_wrap.mat_plot import AutoLabels
 
-from autoarray.plot.structure_plotters import Array2DPlotter
-from autoarray.plot.structure_plotters import Grid2DPlotter
-from autoarray.plot.structure_plotters import MapperPlotter
-from autoarray.plot.structure_plotters import YX1DPlotter
-from autoarray.plot.inversion_plotters import InversionPlotter
-from autoarray.plot.imaging_plotters import ImagingPlotter
-from autoarray.plot.interferometer_plotters import InterferometerPlotter
-from autoarray.plot.fit_imaging_plotters import FitImagingPlotter
-from autoarray.plot.fit_interferometer_plotters import FitInterferometerPlotter
+from autoarray.structures.plot.structure_plotters import Array2DPlotter
+from autoarray.structures.plot.structure_plotters import Grid2DPlotter
+from autoarray.structures.plot.structure_plotters import YX1DPlotter
+from autoarray.inversion.plot.mapper_plotters import MapperPlotter
+from autoarray.inversion.plot.inversion_plotters import InversionPlotter
+from autoarray.dataset.plot.imaging_plotters import ImagingPlotter
+from autoarray.dataset.plot.interferometer_plotters import InterferometerPlotter
+from autoarray.fit.plot.fit_imaging_plotters import FitImagingPlotter
+from autoarray.fit.plot.fit_interferometer_plotters import FitInterferometerPlotter
 
 from autoarray.plot.multi_plotters import MultiFigurePlotter
 from autoarray.plot.multi_plotters import MultiYX1DPlotter
