+++ test_autoarray/dataset/plot/test_imaging_plotters.py
@@ -120,7 +120,7 @@ def test__output_as_fits__correct_output_format(
     imaging_plotter.figures_2d(image=True, psf=True, absolute_signal_to_noise_map=True)
 
     image_from_plot = aa.util.array_2d.numpy_array_2d_from_fits(
-        file_path=path.join(plot_path, "image.fits"), hdu=0
+        file_path=path.join(plot_path, "image_2d.fits"), hdu=0
     )
 
     assert image_from_plot.shape == (7, 7)
